move IrcChannelListing to a separate file
