+++ src/event/irc/EventIrcChatListing.cpp
@@ -7,37 +7,6 @@ UUID EventIrcChatListing::getEventUuid() const {
     return this->uuid;
 }
 
-
-IrcChannelListing::IrcChannelListing(const std::string& channelName,
-                                     const std::string& channelTopic,
-                                     bool disabled)
-    : channelName{channelName}
-    , channelTopic{channelTopic}
-    , disabled{disabled}
-{
-}
-
-void IrcChannelListing::addUser(const std::string& nick, const std::string& mode) {
-    users.emplace_back(nick, mode);
-}
-
-const std::list<IrcChannelUser>& IrcChannelListing::getUsers() const {
-    return users;
-}
-
-
-std::string IrcChannelListing::getChannelName() const {
-    return channelName;
-}
-
-std::string IrcChannelListing::getChannelTopic() const {
-    return channelTopic;
-}
-
-bool IrcChannelListing::getDisabled() const {
-    return disabled;
-}
-
 IrcServerListing::IrcServerListing(std::string activeNick,
                                    size_t serverId,
                                    const std::string& serverName)
