+++ src/event/irc/EventIrcChatListing.hpp
@@ -6,22 +6,6 @@
 #include <list>
 
 
-class IrcChannelListing {
-    std::string channelName;
-    std::string channelTopic;
-    std::list<IrcChannelUser> users;
-    bool disabled;
-public:
-    IrcChannelListing(const std::string& channelName,
-                      const std::string& channelTopic,
-                      bool disabled);
-
-    void addUser(const std::string& nick, const std::string& mode);
-    const std::list<IrcChannelUser>& getUsers() const;
-    std::string getChannelName() const;
-    std::string getChannelTopic() const;
-    bool getDisabled() const;
-};
 
 class IrcServerListing {
     std::string activeNick;
